+++ bt_selector_sequential.c
@@ -247,7 +247,7 @@ int bt_sequential_selector_poll_best_piece(
 
     while (0 < heap_count(pr->p_candidates))
     {
-        piece_idx = (int)heap_poll(pr->p_candidates) - 1;
+        piece_idx = (unsigned long int)heap_poll(pr->p_candidates) - 1;
 
         if (!(hashmap_get(me->p_polled, (void *) (long) piece_idx + 1)))
         {
